* fixing bugs in full replication  * partial replication implemented (SERIALIZABILITY)  * debugging is needed!! (change the invocetion context interface)

+++ core/src/main/java/org/infinispan/distribution/ch/AbstractWheelConsistentHash.java
@@ -249,7 +249,7 @@ public abstract class AbstractWheelConsistentHash extends AbstractConsistentHash
     public ReplGroup getGroupFor(Object key, int replCount) {
         List<Address> addrs = locate(key, replCount);
         List<Address> candidates = new LinkedList<Address>(getRealAddresses(new HashSet(positions.values())));
-        long id = candidates.indexOf(addrs.get(0));
+        int id = candidates.indexOf(addrs.get(0));
         return new ReplGroup(id, addrs);
     }
 }
